+++ pulsar/apps/wsgi/html.py
@@ -1,4 +1,9 @@
+from pulsar.utils.structures import recursive_update
+
 HTML_VISITORS = {}
+
+
+__all__ = ['HtmlVisitor']
 
 
 def html_visitor(tag):
@@ -8,21 +13,33 @@ def html_visitor(tag):
 class HtmlType(type):
     
     def __new__(cls, name, bases, attrs):
+        abstract = attrs.pop('abstract', False)
         new_class = super(HtmlType,cls).__new__(cls, name, bases, attrs)
-        HTML_VISITORS[name.lower()] = new_class()
+        if not abstract:
+            HTML_VISITORS[name.lower()] = new_class()
         return new_class
-
-class Base(object):
+    
+    
+class HtmlVisitor(HtmlType('HtmlVisitorBase', (object,), {'abstract': True})):
+    '''A visitor class for :class:`Html`.'''
+    abstract = True
     
     def get_form_value(self, html):
         return html.attr('value')
     
     def set_form_value(self, html, value):
         html.attr('value', value)
-    
 
-base = Base()
-HtmlVisitor = HtmlType('HtmlVisitor', (Base,), {})
+    def add_data(self, html, key, value):
+        data = html._data
+        if key in data and isinstance(value, Mapping):
+            target = data[key]
+            if isinstance(target, Mapping):
+                return recursive_update(target, value)
+        data[key] = value
+        
+
+base = HtmlVisitor()
     
 class Textarea(HtmlVisitor):